Move create_tool_analysis into `tool_manager.utils`
